Added data struct for rhs of rule
Since there are sort of two cases, where the rhs only has one element or multiple.
The case where there are multiple there needs to be a vector where each index of the element is in order
they appear. To do this effectively recursively I decided to overload the parse_rhs() method to include one
where an address of a vector for rhs is given so that each time a new element occurs the vector gets pushed back.
This then returns a pointer and after all recursive calls get returned this is dereferenced and a copy gets made.

This may result in some funny errors if I'm not careful. Will need to test and keep my eyes peeled.

Will need to do the same with Terminals rules, however additional checks may be applied.