Refactored RHS logic
"Fail to plan, plan to fail"

Knew this ahead of time and continued on arrogantly not looking towards the future.
Realized in order to remove useless symbols from grammar in any way that made sense I need
to refer to elements not by their index in universe but just pointers to those elements.
Code has been refactored. Documentation must be updated later.